requirements.txt
python-telegram-bot[job-queue]==20.8
websockets==12.0
aiohttp==3.9.5
python-dotenv==1.0.1
requests==2.32.3
apscheduler==3.10.4
httpx==0.26.0
